update delete user

+++ src/app/nextui/user/page.tsx
@@ -4,8 +4,16 @@ import { SubmitButton } from "@/components/uinext/SubmitButton";
 import UserModel from "@/db/mongodb/models/user";
 import { getAllUsers } from "@/data/user";
 import AddUserForm from "@/app/nextui/user/AddUserForm";
+import { cookies } from "next/headers";
+import { USER_TOKEN } from "@/lib/constants";
+import { decryptUserToken } from "@/lib/auth";
 
 export default async function Page() {
+  // decrypt user info from token
+  const cookieStore = cookies();
+  const token = cookieStore.get(USER_TOKEN)?.value!;
+  const userPayload = decryptUserToken(token);
+  const { userId, userName } = userPayload;
   const users = await getAllUsers();
 
   return (
@@ -21,13 +29,16 @@ export default async function Page() {
                 className="flex items-center w-full justify-between"
                 action={async () => {
                   "use server";
+                  if (user.id === userId) {
+                    return;
+                  }
                   await UserModel.findByIdAndDelete(user.id);
                   revalidatePath("/nextui/user", "page");
                 }}
               >
                 <div>{user.name}</div>
                 <div>
-                  <SubmitButton text="Delete" />
+                  <SubmitButton disable={user.id === userId} text="Delete" />
                 </div>
               </form>
             </div>
